% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpcp}
\alias{tpcp}
\title{Specific isobaric heat capacity vs temperature and pressure}
\usage{
tpcp(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
Specific isobaric heat capacity, [\eqn{kJ/kg/K}]
}
\description{
Specific isobaric heat capacity vs temperature and pressure
}
