% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpcv}
\alias{tpcv}
\title{Specific isochoric heat capacity vs temperature and pressure}
\usage{
tpcv(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
Specific isochoric heat capacity, [\eqn{kJ/kg/K}]
}
\description{
Specific isochoric heat capacity vs temperature and pressure
}
