% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpe}
\alias{tpe}
\title{Cubic expansion coefficient vs temperature and pressure}
\usage{
tpe(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
cubic expansion coefficient, [\eqn{1/K}]
}
\description{
Cubic expansion coefficient vs temperature and pressure
}
