% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tph}
\alias{tph}
\title{Specific enthalpy vs temperature and pressure}
\usage{
tph(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
specific enthalpy, [\eqn{kJ/kg/K}]
}
\description{
Specific enthalpy vs temperature and pressure
}
