% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpi}
\alias{tpi}
\title{Isothermal compressibility vs temperature and pressure}
\usage{
tpi(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
Isothermal compressibility, [\eqn{1/MPa}]
}
\description{
Isothermal compressibility vs temperature and pressure
}
