% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tps}
\alias{tps}
\title{Specific entropy vs temperature and pressure}
\usage{
tps(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
specific entropy, [\eqn{kJ/kg/K}]
}
\description{
Specific entropy vs temperature and pressure
}
