% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpu}
\alias{tpu}
\title{Specific internal energy vs temperature and pressure}
\usage{
tpu(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
specific internal energy, [\eqn{kJ/kg}]
}
\description{
Specific internal energy vs temperature and pressure
}
