% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpv}
\alias{tpv}
\title{Specific volume vs temperature and pressure}
\usage{
tpv(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
specific volume, [\eqn{m^3/kg}]
}
\description{
Specific volume vs temperature and pressure
}
