% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpw}
\alias{tpw}
\title{Speed of sound vs temperature and pressure}
\usage{
tpw(T, p)
}
\arguments{
\item{T}{absolute temperature, [\emph{K}]}

\item{p}{absolute pressure, [\emph{MPa}]}
}
\value{
Speed of sound, [\eqn{m/s}]
}
\description{
Speed of sound vs temperature and pressure
}
