% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tpv}
\alias{tpv}
\title{Specific volume vs temperature and pressure}
\usage{
tpv(T, p)
}
\arguments{
\item{T}{absolute temperature in validity range 273.15 -- 2273.15, [\emph{K}].
Type: \code{NumericVector}}

\item{p}{absolute pressure in validity range 0 -- 100 \emph{MPa}, [\emph{MPa}].
Type: \code{NumericVector}}
}
\value{
specific volume, [\eqn{m^3/kg}], or the next error code:
  \describe{
      \item{\code{-10}}{
        Fail to determine region in temperature-pressure space into which the
        entire range of validity of \emph{IAPWS R7-97}
        is divided. Possibly values of one or both arguments are
        out of bounds or in undeterminable subspace.
      }
      \item{\code{-20}}{
        Fail to determine subregion inside \emph{Region 3} in accordance
        with \emph{IAPWS SR5-05}. Possibly values of one or both arguments
        are in undeterminable subspace or near singularity.
      }
  }
  Type: \code{NumericVector}
}
\description{
Calculate \emph{specific volume} as a function of \emph{temperature} and
  \emph{pressure} in accordance with \emph{IAPWS R7-97} formulation.
}
\details{
Vectors \code{T} and \code{p} both must have the same length. Default
  \href{https://cran.r-project.org/doc/manuals/r-release/R-intro.html#Vector-arithmetic}{recycling rules}
  are not applicable in this context.

  In current version numerical consistency of the specific volume
  in \emph{Region 3} of \emph{IAPWS R7-97} is considered sufficient enough
  for most applications. That is why no additional iteration procedures are
  used for reaching higher accuracy in this region.
}
\examples{
 t <- c(300,300,500,300,700,700,1500,1500,2000,0)  # [K]
 p <- c(3,80,3,.35e-2,.35e-2,30,.5,30,30,0)  # [MPa]
 tpv(t, p)

}
\references{
\itemize{
  \item \href{http://www.iapws.org/relguide/IF97-Rev.pdf}{IAPWS R7-97 (2012)},
  \emph{Revised Release on the IAPWS Industrial Formulation 1997 for the
  Thermodynamic Properties of Water and Steam}. August 2007.

  \item \href{http://www.iapws.org/relguide/Supp-VPT3-2016.pdf}{IAPWS SR5-05 (2016)},
  \emph{Revised Supplementary Release on Backward Equations for Specific
  Volume as a Function of Pressure and Temperature v(p,T) for Region 3 of the
  IAPWS Industrial Formulation 1997 for the Thermodynamic Properties of Water and
  Steam}. June 2014.
 }
}
\seealso{
Other properties in temperature-pressure space: 
\code{\link{tpcp}()},
\code{\link{tpcv}()},
\code{\link{tpe}()},
\code{\link{tph}()},
\code{\link{tpi}()},
\code{\link{tpr}()},
\code{\link{tps}()},
\code{\link{tpu}()},
\code{\link{tpw}()}
}
\concept{properties in temperature-pressure space}
